@@ -14,56 +14,61 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.10",
-    "@angular/cdk": "^5.2.4",
-    "@angular/common": "^5.2.10",
-    "@angular/compiler": "^5.2.10",
-    "@angular/core": "^5.2.10",
-    "@angular/flex-layout": "^5.0.0-beta.14",
-    "@angular/forms": "^5.2.10",
-    "@angular/http": "^5.2.10",
-    "@angular/material": "^5.2.4",
-    "@angular/platform-browser": "^5.2.10",
-    "@angular/platform-browser-dynamic": "^5.2.10",
-    "@angular/router": "^5.2.10",
-    "@ngrx/effects": "^5.2.0",
-    "@ngrx/store": "^5.2.0",
-    "@ngrx/store-devtools": "^5.2.0",
-    "angularfire2": "^5.0.0-rc.5-next",
+    "@angular-devkit/schematics": "^10.0.5",
+    "@angular/animations": "10.0.8",
+    "@angular/cdk": "^10.1.2",
+    "@angular/common": "10.0.8",
+    "@angular/compiler": "10.0.8",
+    "@angular/core": "10.0.8",
+    "@angular/fire": "^6.0.2",
+    "@angular/flex-layout": "^10.0.0-beta.32",
+    "@angular/forms": "10.0.8",
+    "@angular/material": "^10.1.2",
+    "@angular/platform-browser": "10.0.8",
+    "@angular/platform-browser-dynamic": "10.0.8",
+    "@angular/router": "10.0.8",
+    "@ngrx/effects": "^9.2.0",
+    "@ngrx/store": "^9.2.0",
+    "@ngrx/store-devtools": "^9.2.0",
     "brace": "^0.11.1",
-    "core-js": "^2.5.5",
-    "firebase": "^4.12.1",
+    "core-js": "^3.6.5",
+    "firebase": "^7.17.1",
+    "grpc": "^1.24.3",
     "hammerjs": "^2.0.8",
-    "ng2-ace-editor": "^0.3.4",
+    "ng2-ace-editor": "^0.3.9",
     "ngrx-store-freeze": "^0.2.2",
+    "node-gyp": "^7.0.0",
+    "node-pre-gyp": "^0.15.0",
     "reselect": "^3.0.1",
-    "rxjs": "^5.5.10",
+    "rxjs": "^6.6.2",
     "socket.io-client": "^2.1.0",
-    "zone.js": "^0.8.26"
+    "tslib": "^1.9.0",
+    "zone.js": "~0.10.3"
   },
   "devDependencies": {
-    "@angular/cli": "^1.7.4",
-    "@angular/compiler-cli": "^5.2.10",
-    "@angular/language-service": "^5.2.10",
+    "@angular-devkit/build-angular": "~0.1000.5",
+    "@angular/cli": "^10.0.5",
+    "@angular/compiler-cli": "10.0.8",
+    "@angular/language-service": "10.0.8",
     "@types/jasmine": "~2.8.6",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "^9.6.5",
     "@types/socket.io-client": "^1.4.32",
-    "codelyzer": "^4.2.1",
-    "concurrently": "^3.5.1",
+    "codelyzer": "^6.0.0",
+    "concurrently": "^5.2.0",
     "jasmine-core": "~3.1.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "json-server": "^0.12.1",
-    "karma": "~2.0.0",
+    "json-server": "^0.16.1",
+    "karma": "^5.1.1",
     "karma-chrome-launcher": "~2.2.0",
     "karma-cli": "~1.0.1",
     "karma-coverage-istanbul-reporter": "^1.4.2",
     "karma-jasmine": "^1.1.1",
     "karma-jasmine-html-reporter": "^1.0.0",
-    "protractor": "~5.3.1",
-    "socket.io": "^2.1.0",
-    "ts-node": "~5.0.1",
+    "protractor": "^7.0.0",
+    "socket.io": "^2.3.0",
+    "ts-node": "~8.10.2",
     "tslint": "~5.9.1",
-    "typescript": "^2.6.2"
+    "typescript": "3.9.7"
   }
 }
